requirements.txt
mysql-connector-python
python-barcode
pyzbar
customtkinter
matplotlib
CTkTable
pywinstyles